requirements.txt
Flask==2.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0
WTForms==3.0
Werkzeug==2.3.7
Pillow==10.0
email-validator==2.0
python-dotenv==1.0
Flask-Admin==1.6.1 
gunicorn
